src/requirements.txt
taipy==3.0.0.dev3
rembg
pillow